Add global appbar
