+++ frontend/components/change-password.tsx
@@ -92,65 +92,63 @@ const ChangePasswordPage = () => {
   };
   return (
     <Container sx={{ height: '100%' }}>
-      <DefaultLayout>
-        <div className="change-password-page">
-          <h2>Change Password</h2>
-          <form onSubmit={handlePasswordChange}>
-            <TextField
-              error={oldPasswordError}
-              helperText={oldPasswordError ? 'Old Password does not match' : ''}
-              style={{ width: '200px', margin: '5px' }}
-              type="password"
-              label="Old Password"
-              value={oldPassword}
-              name="oldPassword"
-              onChange={(e) => setOldPassword(e.target.value)}
-              onKeyUp={handleValidation}
-              variant="filled"
-            />
-            <br />
-            <TextField
-              error={passwordError.length > 0}
-              helperText={passwordError ?? ''}
-              style={{ width: '200px', margin: '5px', marginTop: '30px' }}
-              type="password"
-              label="New Password"
-              variant="outlined"
-              name="newPassword"
-              value={newPassword}
-              onChange={(e) => setNewPassword(e.target.value)}
-              onKeyUp={handleValidation}
-            />
-            <br />
-            <TextField
-              error={confirmPasswordError}
-              helperText={confirmPasswordError ? 'Password does not match' : ''}
-              style={{ width: '200px', margin: '5px', marginTop: '15px' }}
-              type="password"
-              label="Confirm New Password"
-              variant="outlined"
-              name="confirmNewPassword"
-              value={confirmNewPassword}
-              onChange={(e) => setConfirmNewPassword(e.target.value)}
-              onKeyUp={handleValidation}
-            />
-            <br />
-            <Button
-              disabled={
-                confirmPasswordError ||
-                passwordError !== '' ||
-                newPassword === '' ||
-                confirmNewPassword === ''
-              }
-              variant="contained"
-              color="primary"
-              type="submit"
-            >
-              Change Password
-            </Button>
-          </form>
-        </div>
-      </DefaultLayout>
+      <div className="change-password-page">
+        <h2>Change Password</h2>
+        <form onSubmit={handlePasswordChange}>
+          <TextField
+            error={oldPasswordError}
+            helperText={oldPasswordError ? 'Old Password does not match' : ''}
+            style={{ width: '200px', margin: '5px' }}
+            type="password"
+            label="Old Password"
+            value={oldPassword}
+            name="oldPassword"
+            onChange={(e) => setOldPassword(e.target.value)}
+            onKeyUp={handleValidation}
+            variant="filled"
+          />
+          <br />
+          <TextField
+            error={passwordError.length > 0}
+            helperText={passwordError ?? ''}
+            style={{ width: '200px', margin: '5px', marginTop: '30px' }}
+            type="password"
+            label="New Password"
+            variant="outlined"
+            name="newPassword"
+            value={newPassword}
+            onChange={(e) => setNewPassword(e.target.value)}
+            onKeyUp={handleValidation}
+          />
+          <br />
+          <TextField
+            error={confirmPasswordError}
+            helperText={confirmPasswordError ? 'Password does not match' : ''}
+            style={{ width: '200px', margin: '5px', marginTop: '15px' }}
+            type="password"
+            label="Confirm New Password"
+            variant="outlined"
+            name="confirmNewPassword"
+            value={confirmNewPassword}
+            onChange={(e) => setConfirmNewPassword(e.target.value)}
+            onKeyUp={handleValidation}
+          />
+          <br />
+          <Button
+            disabled={
+              confirmPasswordError ||
+              passwordError !== '' ||
+              newPassword === '' ||
+              confirmNewPassword === ''
+            }
+            variant="contained"
+            color="primary"
+            type="submit"
+          >
+            Change Password
+          </Button>
+        </form>
+      </div>
     </Container>
   );
 };
