+++ frontend/layouts/DefaultLayout.tsx
@@ -1,6 +1,5 @@
+import AppBar from '@/components/defaultLayout/AppBar';
 import useUserStore from '@/lib/store';
-import { Container } from '@mui/material';
-import Head from 'next/head';
 import React from 'react';
 
 interface DefaultLayoutProps {
@@ -11,15 +10,11 @@ function DefaultLayout({ children }: DefaultLayoutProps) {
   const user = useUserStore((state) => state.user);
 
   return (
-    <Container>
-      <Head>
-        <title>Leet Deez Nutz</title>
-        <meta name="viewport" content="initial-scale=1.0, width=device-width" />
-      </Head>
+    <>
+      <AppBar />
 
-      <div>{user?.username ?? 'Not logged in'}</div>
       {children}
-    </Container>
+    </>
   );
 }
