+++ frontend/pages/_app.tsx
@@ -3,6 +3,7 @@ import type { AppProps } from 'next/app';
 import { useEffect } from 'react';
 import useUserStore from '@/lib/store';
 import { getJwtCookie } from '@/lib/cookies';
+import Head from 'next/head';
 
 function MyApp({ Component, pageProps }: AppProps) {
   const documentIsReady = typeof window !== 'undefined';
@@ -20,7 +21,15 @@ function MyApp({ Component, pageProps }: AppProps) {
     updateUser(cookie);
   }, [documentIsReady]);
 
-  return <Component {...pageProps} />;
+  return (
+    <>
+      <Head>
+        <title>Leet Deez Nutz</title>
+        <meta name="viewport" content="initial-scale=1.0, width=device-width" />
+      </Head>
+      <Component {...pageProps} />;
+    </>
+  );
 }
 
 export default MyApp;
