+++ frontend/pages/learning-pathway.tsx
@@ -1,0 +1,7 @@
+import DefaultLayout from '@/layouts/DefaultLayout';
+
+const LearningPathway = () => {
+  return <DefaultLayout>hi</DefaultLayout>;
+};
+
+export default LearningPathway;
